+++ src/app/layout/index.tsx
@@ -1,7 +1,7 @@
 import {Outlet, useLocation, useParams} from 'react-router-dom';
 import {Helmet} from 'react-helmet-async';
-import {getLayoutState} from './utils';
 import {Footer, Header} from '@/widgets';
+import {getLayoutState} from './utils';
 
 export const Layout = () => {
   const {pathname} = useLocation();
