+++ src/pages/login/model.ts
@@ -1,8 +1,8 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
-import {PromiseProps, UserProps} from '../../shared/types';
-import {redirectToRoute, token} from '../../shared/utils';
-import {Path} from '../../shared/config';
-import {checkAuthAction} from "@/shared/api";
+import {PromiseProps, UserProps} from '@/shared/types';
+import {redirectToRoute, token} from '@/shared/utils';
+import {Path} from '@/shared/config';
+import {checkAuthAction} from '@/shared/api';
 
 type AuthData = {
   email: string;
