+++ src/pages/login/ui.tsx
@@ -1,9 +1,9 @@
 import {ChangeEvent, FormEvent, memo, useState} from 'react';
 import {Link, Navigate} from 'react-router-dom';
-import {useAppDispatch, useAppSelector} from '../../app/app-store';
-import {Path} from '../../shared/config';
-import {formFields} from "./const";
-import {getAuthCheckedStatus, VisuallyHidden} from '../../shared/utils';
+import {useAppDispatch, useAppSelector} from '@/app/app-store';
+import {Path} from '@/shared/config';
+import {getAuthCheckedStatus, VisuallyHidden} from '@/shared/utils';
+import {formFields} from './const';
 import {loginAction} from './model';
 
 export const Login = memo(() => {
