+++ src/pages/main/ui/places/type.ts
@@ -1,5 +1,5 @@
-import {PreviewOfferProps} from '../../../../shared/types';
-import {CityName} from '../../../../shared/config';
+import {PreviewOfferProps} from '@/shared/types';
+import {CityName} from '@/shared/config';
 
 export type PlacesProps = {
   numberPlacesToStay: number;
