+++ src/pages/offer/api.ts
@@ -1,6 +1,6 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
+import {ExtraType, OfferProp, PreviewOfferProps} from '@/shared/types';
 import {CommentsProps} from './type';
-import {ExtraType, OfferProp, PreviewOfferProps} from '../../shared/types';
 
 type PostCommentsProps = {
   id: string;
