+++ src/pages/offer/index.tsx
@@ -1,12 +1,12 @@
 import {useEffect} from 'react';
 import {useParams} from 'react-router-dom';
-import {useAppDispatch, useAppSelector} from '../../app/app-store';
-import {getAuthCheckedStatus} from '../../shared/utils';
-import {OfferProp} from '../../shared/types';
-import {Loader, Map, Rating, User} from '../../shared';
-import {Bookmark} from '../../feature';
-import {PlaceCard} from '../../entities';
-import {MAX_NEAR_PLACES, MAX_COMMENTS} from "./const";
+import {useAppDispatch, useAppSelector} from '@/app/app-store';
+import {getAuthCheckedStatus} from '@/shared/utils';
+import {OfferProp} from '@/shared/types';
+import {Loader, Map, Rating, User} from '@/shared';
+import {Bookmark} from '@/feature';
+import {PlaceCard} from '@/entities';
+import {MAX_NEAR_PLACES, MAX_COMMENTS} from './const';
 import {fetchCommentsAction, fetchCurrentOfferAction, fetchNearbyAction} from './api';
 import {getComments, getNearPlaces, getOffer} from './model';
 import {ReviewForm, Reviews} from './ui';
