+++ src/pages/offer/model.ts
@@ -1,9 +1,9 @@
 import {createSelector, createSlice} from '@reduxjs/toolkit';
+import {Status} from '@/shared/config';
+import {PreviewOfferProps, OfferProp, TypeState} from '@/shared/types';
+import {postFavoriteStatusAction} from '@/shared/api';
 import {CommentsProps} from './type';
 import {fetchCommentsAction, fetchCurrentOfferAction, fetchNearbyAction, postReviewAction} from './api';
-import {Status} from '../../shared/config';
-import {PreviewOfferProps, OfferProp, TypeState} from '../../shared/types';
-import {postFavoriteStatusAction} from '../../shared/api';
 
 type InitialStateProps = {
   statusOffer: null | Status;
