+++ src/shared/api/create-api/index.ts
@@ -1,9 +1,9 @@
 import axios, {AxiosInstance, InternalAxiosRequestConfig, AxiosError} from 'axios';
 import {StatusCodes} from 'http-status-codes';
 import { toast } from 'react-toastify';
-import {token} from '../../utils/token';
-import {browserHistory} from '../../utils/browser-history';
-import {Path} from '../../config';
+import {token} from '@/shared/utils/token';
+import {Path} from '@/shared/config';
+import {browserHistory} from '@/shared/utils/browser-history';
 
 type DetailMessageType = {
   type: string;
