+++ src/shared/api/fetch-favorite-action/index.ts
@@ -1,5 +1,5 @@
-import {createAsyncThunk} from "@reduxjs/toolkit";
-import {ExtraType, PreviewOfferProps} from "@/shared/types";
+import {createAsyncThunk} from '@reduxjs/toolkit';
+import {ExtraType, PreviewOfferProps} from '@/shared/types';
 
 export const fetchFavoriteAction = createAsyncThunk<PreviewOfferProps[], undefined, ExtraType>(
   'favorite/fetchOffers',
