+++ src/shared/utils/redirect-to-route/index.ts
@@ -1,4 +1,4 @@
 import {createAction} from '@reduxjs/toolkit';
-import {Path} from '../../config';
+import {Path} from '@/shared/config';
 
 export const redirectToRoute = createAction<Path>('utils/redirectToRoute');
