+++ src/widgets/footer/index.tsx
@@ -1,6 +1,6 @@
 import {memo} from 'react';
 import {Link} from 'react-router-dom';
-import {Path} from '../../shared/config';
+import {Path} from '@/shared/config';
 
 export const Footer = memo(() => (
   <footer className="footer container">
